requirements.txt
aiofiles==0.7.0
aiohttp==3.8.0
aiohttp-swagger==1.0.16
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.0
asynctest==0.13.0
asyncua==0.9.91
attrs==21.2.0
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
frozenlist==1.2.0
idna==3.3
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
six==1.16.0
sortedcontainers==2.4.0
typing-extensions==3.10.0.2
yarl==1.7.2
